The lead concern keeps to its own file
Adding the lane pushed lanes.go and worklist.go past the 500-line cap, so the
lead's declarations, its classifier and the read that fills it move together
into lead.go. They are one concept and they read one another; splitting them
across three files is what the cap exists to prevent.

subjectLead was already spelled in bands.go. Using it rather than declaring a
second is the same rule the lane itself follows: one answer to what a lead row
points at.
